package: describe the relationship to style-dictionary
style-dictionary is a peer dependency, a vehicle used to "express the
compatibility of your package with a host tool or library"[0].

Expressing the relationship make sense both conceptually (there is no
point in people using style-dictionary-format-json-schema without using
style-dictionary) and also from a practical point of view (as
style-dictionary-format-json-schema uses the style-dictionary API which
might change).

I could not find (did not try hard, though) about the registerFormat()
API of style-dictionary version 1 and if it was compatible but I doubt
there are many users of it any more.

[0]: https://docs.npmjs.com/cli/v6/configuring-npm/package-json#peerdependencies